+++ src/solver/consistency/consistency.rs
@@ -11,33 +11,38 @@ use crate::csp::csp::Csp;
 use crate::csp::prelude::domain::OrdT;
 pub(crate) use crate::solver::consistency::revise::Revise;
 use crate::solver::consistency::scheme::{Scheme};
+use crate::instrumentation::monitor::Monitor;
 
 //TriggerEvent
 
-pub struct Consistency<S, R, T:OrdT>
+pub struct Consistency<M, S, R, T:OrdT>
 where
-    S: Scheme<T, R>,
-    R: Revise<T>,
+    S: Scheme<M, T, R>,
+    R: Revise<M, T>,
+    M: Monitor,
 {
     scheme: S,
     revise: R,
+    monitor: M,
     _phantom: std::marker::PhantomData<T>
 }
 
-impl<S, R, T:OrdT> Consistency<S, R, T>
+impl<M, S, R, T:OrdT> Consistency<M, S, R, T>
 where
-    S: Scheme<T, R>,
-    R: Revise<T>,
+    S: Scheme<M, T, R>,
+    R: Revise<M, T>,
+    M: Monitor
 {
-    pub fn new(scheme: S, revise: R) -> Self {
+    pub fn new(scheme: S, revise: R, monitor: M) -> Self {
         Self {
             scheme,
             revise,
+            monitor,
             _phantom: std::marker::PhantomData
         }
     }
 
     pub fn enforce_consistency(&mut self, csp: &mut Csp<T>, events: Vec<String>) {
-        self.scheme.enforce(csp, events, &mut self.revise);
+        self.scheme.enforce(csp, events, &mut self.revise, &mut self.monitor);
     }
 }